requirements.txt
streamlit~=1.47.0
groq~=0.30.0
gspread~=6.2.1
oauth2client~=4.1.3
requests
beautifulsoup4
google-auth
google-auth-oauthlib
